Move more constructor tests, to_pandas, squeeze
